Combine sprite text to lyric text.
