+++ osu.Framework.Font/Graphics/Sprites/LyricSpriteText.cs
@@ -1,370 +1,597 @@
-﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the MIT Licence.
+// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using osu.Framework.Text;
+using System.Diagnostics;
+using osu.Framework.Allocation;
+using osu.Framework.Bindables;
+using osu.Framework.Extensions.IEnumerableExtensions;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shaders;
+using osu.Framework.Graphics.UserInterface;
+using osu.Framework.IO.Stores;
+using osu.Framework.Layout;
+using osu.Framework.Localisation;
 using osu.Framework.Utils;
+using osu.Framework.Text;
 using osuTK;
 using osuTK.Graphics;
 
 namespace osu.Framework.Graphics.Sprites
 {
-    public class LyricSpriteText : SpriteText, IHasRuby, IHasRomaji, IHasTexture
+    /// <summary>
+    /// A container for simple text rendering purposes. If more complex text rendering is required, use <see cref="TextFlowContainer"/> instead.
+    /// </summary>
+    public partial class LyricSpriteText : Drawable, IHasLineBaseHeight, ITexturedShaderDrawable, IHasFilterTerms, IFillFlowContainer, IHasCurrentValue<string>, IHasRuby, IHasRomaji, IHasTexture
     {
+        private const float default_text_size = 20;
+        private static readonly char[] default_never_fixed_width_characters = { '.', ',', ':', ' ' };
+
+        [Resolved]
+        private FontStore store { get; set; }
+
+        [Resolved]
+        private LocalisationManager localisation { get; set; }
+
+        public IShader TextureShader { get; private set; }
+        public IShader RoundedTextureShader { get; private set; }
+
         public LyricSpriteText()
         {
             Font = new FontUsage(null, 48);
+            current.BindValueChanged(text => Text = text.NewValue);
+
+            AddLayout(charactersCache);
+            AddLayout(parentScreenSpaceCache);
+            AddLayout(localScreenSpaceCache);
+            AddLayout(shadowOffsetCache);
         }
 
-        private PositionText[] rubies;
-
-        public PositionText[] Rubies
+        [BackgroundDependencyLoader]
+        private void load(ShaderManager shaders)
         {
-            get => rubies;
-            set
+            TextureShader = shaders.Load(VertexShaderDescriptor.TEXTURE_2, FragmentShaderDescriptor.TEXTURE);
+            RoundedTextureShader = shaders.Load(VertexShaderDescriptor.TEXTURE_2, FragmentShaderDescriptor.TEXTURE_ROUNDED);
+
+            // Pre-cache the characters in the texture store
+            foreach (var character in displayedText)
             {
-                rubies = filterValidValues(value);
-                Invalidate(Invalidation.All);
+                var unused = store.Get(font.FontName, character) ?? store.Get(null, character);
             }
         }
 
-        private PositionText[] romajies;
+        private string text = string.Empty;
 
-        public PositionText[] Romajies
+        /// <summary>
+        /// Gets or sets the text to be displayed.
+        /// </summary>
+        public string Text
         {
-            get => romajies;
+            get => text;
             set
             {
-                romajies = filterValidValues(value);
-                Invalidate(Invalidation.All);
+                if (text == value)
+                    return;
+
+                text = value;
+
+                current.Value = text;
             }
         }
 
-        private PositionText[] filterValidValues(PositionText[] texts)
+        private readonly BindableWithCurrent<string> current = new BindableWithCurrent<string>();
+
+        public Bindable<string> Current
         {
-            string text = Text;
-            return texts?.Where(positionText => Math.Min(positionText.StartIndex, positionText.EndIndex) >= 0
-                                                && Math.Max(positionText.StartIndex, positionText.EndIndex) <= text.Length
-                                                && positionText.EndIndex > positionText.StartIndex).ToArray();
+            get => current.Current;
+            set => current.Current = value;
         }
 
-        private FontUsage rubyFont = FontUsage.Default;
+        private string displayedText => Text;
+
+        private FontUsage font = FontUsage.Default;
 
         /// <summary>
         /// Contains information on the font used to display the text.
         /// </summary>
-        public FontUsage RubyFont
+        public FontUsage Font
         {
-            get => rubyFont;
+            get => font;
             set
             {
-                rubyFont = value;
-                Invalidate(Invalidation.All);
+                font = value;
+
+                invalidate(true);
+                shadowOffsetCache.Invalidate();
             }
         }
 
-        private FontUsage romajiFont = FontUsage.Default;
+        private bool allowMultiline = true;
 
         /// <summary>
-        /// Contains information on the font used to display the text.
+        /// True if the text should be wrapped if it gets too wide. Note that \n does NOT cause a line break. If you need explicit line breaks, use <see cref="TextFlowContainer"/> instead.
         /// </summary>
-        public FontUsage RomajiFont
+        /// <remarks>
+        /// If enabled, <see cref="Truncate"/> will be disabled.
+        /// </remarks>
+        public bool AllowMultiline
         {
-            get => romajiFont;
+            get => allowMultiline;
             set
             {
-                romajiFont = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private int rubyMargin;
-
-        public int RubyMargin
-        {
-            get => rubyMargin;
-            set
-            {
-                if (rubyMargin == value)
+                if (allowMultiline == value)
                     return;
 
-                rubyMargin = value;
-                Invalidate(Invalidation.All);
+                if (value)
+                    Truncate = false;
+
+                allowMultiline = value;
+
+                invalidate(true);
             }
         }
 
-        private int romajiMargin;
-
-        public int RomajiMargin
-        {
-            get => romajiMargin;
-            set
-            {
-                if (romajiMargin == value)
-                    return;
-
-                romajiMargin = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private Vector2 rubySpacing;
-
-        public Vector2 RubySpacing
-        {
-            get => rubySpacing;
-            set
-            {
-                if (rubySpacing == value)
-                    return;
-
-                rubySpacing = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private Vector2 romajiSpacing;
-
-        public Vector2 RomajiSpacing
-        {
-            get => romajiSpacing;
-            set
-            {
-                if (romajiSpacing == value)
-                    return;
-
-                romajiSpacing = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private ILyricTexture textTexture;
-
-        public ILyricTexture TextTexture
-        {
-            get => textTexture;
-            set
-            {
-                if (textTexture == value)
-                    return;
-
-                textTexture = value;
-                Colour = (textTexture as SolidTexture)?.SolidColor ?? Color4.White;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private ILyricTexture shadowTexture;
-
-        public ILyricTexture ShadowTexture
-        {
-            get => shadowTexture;
-            set
-            {
-                if (shadowTexture == value)
-                    return;
-
-                shadowTexture = value;
-                ShadowColour = (shadowTexture as SolidTexture)?.SolidColor ?? Color4.White;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private ILyricTexture borderTexture;
-
-        public ILyricTexture BorderTexture
-        {
-            get => borderTexture;
-            set
-            {
-                if (borderTexture == value)
-                    return;
-
-                borderTexture = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private LyricTextAlignment rubyAlignment;
-
-        public LyricTextAlignment RubyAlignment
-        {
-            get => rubyAlignment;
-            set
-            {
-                if (rubyAlignment == value)
-                    return;
-
-                rubyAlignment = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private LyricTextAlignment romajiAlignment;
-
-        public LyricTextAlignment RomajiAlignment
-        {
-            get => romajiAlignment;
-            set
-            {
-                if (romajiAlignment == value)
-                    return;
-
-                romajiAlignment = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private float borderRadius;
-
-        public float BorderRadius
-        {
-            get => borderRadius;
-            set
-            {
-                if (borderRadius == value)
-                    return;
-
-                borderRadius = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        private bool border;
-
-        public bool Border
-        {
-            get => border;
-            set
-            {
-                if (border == value)
-                    return;
-
-                border = value;
-                Invalidate(Invalidation.All);
-            }
-        }
-
-        public new Vector2 ShadowOffset
-        {
-            get => base.ShadowOffset * Font.Size;
-            set => base.ShadowOffset = value / Font.Size;
-        }
-
-        // Store characters without ruby and romaji
-        protected TextBuilderGlyph[] Characters;
+        private bool shadow;
 
         /// <summary>
-        /// Creates a <see cref="TextBuilder"/> to generate the character layout for this <see cref="SpriteText"/>.
+        /// True if a shadow should be displayed around the text.
+        /// </summary>
+        public bool Shadow
+        {
+            get => shadow;
+            set
+            {
+                if (shadow == value)
+                    return;
+
+                shadow = value;
+
+                Invalidate(Invalidation.DrawNode);
+            }
+        }
+
+        private Color4 shadowColour = new Color4(0, 0, 0, 0.2f);
+
+        /// <summary>
+        /// The colour of the shadow displayed around the text. A shadow will only be displayed if the <see cref="Shadow"/> property is set to true.
+        /// </summary>
+        public Color4 ShadowColour
+        {
+            get => shadowColour;
+            set
+            {
+                if (shadowColour == value)
+                    return;
+
+                shadowColour = value;
+
+                Invalidate(Invalidation.DrawNode);
+            }
+        }
+
+        private Vector2 shadowOffset = new Vector2(0, 0.06f);
+
+        /// <summary>
+        /// The offset of the shadow displayed around the text. A shadow will only be displayed if the <see cref="Shadow"/> property is set to true.
+        /// </summary>
+        public Vector2 ShadowOffset
+        {
+            get => shadowOffset;
+            set
+            {
+                // Need to change size to percentage of font size.
+                var percentagedSize = value / Font.Size;
+                if (shadowOffset == percentagedSize)
+                    return;
+
+                shadowOffset = percentagedSize;
+
+                invalidate(true);
+                shadowOffsetCache.Invalidate();
+            }
+        }
+
+        private bool useFullGlyphHeight = true;
+
+        /// <summary>
+        /// True if the <see cref="LyricSpriteText"/>'s vertical size should be equal to <see cref="FontUsage.Size"/>  (the full height) or precisely the size of used characters.
+        /// Set to false to allow better centering of individual characters/numerals/etc.
+        /// </summary>
+        public bool UseFullGlyphHeight
+        {
+            get => useFullGlyphHeight;
+            set
+            {
+                if (useFullGlyphHeight == value)
+                    return;
+
+                useFullGlyphHeight = value;
+
+                invalidate(true);
+            }
+        }
+
+        private bool truncate;
+
+        /// <summary>
+        /// If true, text should be truncated when it exceeds the <see cref="Drawable.DrawWidth"/> of this <see cref="LyricSpriteText"/>.
+        /// </summary>
+        /// <remarks>
+        /// Has no effect if no <see cref="Width"/> or custom sizing is set.
+        /// If enabled, <see cref="AllowMultiline"/> will be disabled.
+        /// </remarks>
+        public bool Truncate
+        {
+            get => truncate;
+            set
+            {
+                if (truncate == value) return;
+
+                if (value)
+                    AllowMultiline = false;
+
+                truncate = value;
+                invalidate(true);
+            }
+        }
+
+        private string ellipsisString = "…";
+
+        /// <summary>
+        /// When <see cref="Truncate"/> is enabled, this decides what string is used to signify that truncation has occured.
+        /// Defaults to "…".
+        /// </summary>
+        public string EllipsisString
+        {
+            get => ellipsisString;
+            set
+            {
+                if (ellipsisString == value) return;
+
+                ellipsisString = value;
+                invalidate(true);
+            }
+        }
+
+        private bool requiresAutoSizedWidth => explicitWidth == null && (RelativeSizeAxes & Axes.X) == 0;
+
+        private bool requiresAutoSizedHeight => explicitHeight == null && (RelativeSizeAxes & Axes.Y) == 0;
+
+        private float? explicitWidth;
+
+        /// <summary>
+        /// Gets or sets the width of this <see cref="LyricSpriteText"/>. The <see cref="LyricSpriteText"/> will maintain this width when set.
+        /// </summary>
+        public override float Width
+        {
+            get
+            {
+                if (requiresAutoSizedWidth)
+                    computeCharacters();
+                return base.Width;
+            }
+            set
+            {
+                if (explicitWidth == value)
+                    return;
+
+                base.Width = value;
+                explicitWidth = value;
+
+                invalidate(true);
+            }
+        }
+
+        private float maxWidth = float.PositiveInfinity;
+
+        /// <summary>
+        /// The maximum width of this <see cref="LyricSpriteText"/>. Affects both auto and fixed sizing modes.
+        /// </summary>
+        /// <remarks>
+        /// This becomes a relative value if this <see cref="LyricSpriteText"/> is relatively-sized on the X-axis.
+        /// </remarks>
+        public float MaxWidth
+        {
+            get => maxWidth;
+            set
+            {
+                if (maxWidth == value)
+                    return;
+
+                maxWidth = value;
+                invalidate(true);
+            }
+        }
+
+        private float? explicitHeight;
+
+        /// <summary>
+        /// Gets or sets the height of this <see cref="LyricSpriteText"/>. The <see cref="LyricSpriteText"/> will maintain this height when set.
+        /// </summary>
+        public override float Height
+        {
+            get
+            {
+                if (requiresAutoSizedHeight)
+                    computeCharacters();
+                return base.Height;
+            }
+            set
+            {
+                if (explicitHeight == value)
+                    return;
+
+                base.Height = value;
+                explicitHeight = value;
+
+                invalidate(true);
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the size of this <see cref="LyricSpriteText"/>. The <see cref="LyricSpriteText"/> will maintain this size when set.
+        /// </summary>
+        public override Vector2 Size
+        {
+            get
+            {
+                if (requiresAutoSizedWidth || requiresAutoSizedHeight)
+                    computeCharacters();
+                return base.Size;
+            }
+            set
+            {
+                Width = value.X;
+                Height = value.Y;
+            }
+        }
+
+        private Vector2 spacing;
+
+        /// <summary>
+        /// Gets or sets the spacing between characters of this <see cref="LyricSpriteText"/>.
+        /// </summary>
+        public Vector2 Spacing
+        {
+            get => spacing;
+            set
+            {
+                if (spacing == value)
+                    return;
+
+                spacing = value;
+
+                invalidate(true);
+            }
+        }
+
+        private MarginPadding padding;
+
+        /// <summary>
+        /// Shrinks the space which may be occupied by characters of this <see cref="LyricSpriteText"/> by the specified amount on each side.
+        /// </summary>
+        public MarginPadding Padding
+        {
+            get => padding;
+            set
+            {
+                if (padding.Equals(value))
+                    return;
+
+                if (!Validation.IsFinite(value)) throw new ArgumentException($@"{nameof(Padding)} must be finite, but is {value}.");
+
+                padding = value;
+
+                invalidate(true);
+            }
+        }
+
+        public override bool IsPresent => base.IsPresent && (AlwaysPresent || !string.IsNullOrEmpty(displayedText));
+
+        #region Characters
+
+        private readonly LayoutValue charactersCache = new LayoutValue(Invalidation.DrawSize | Invalidation.Presence, InvalidationSource.Parent);
+
+        /// <summary>
+        /// Glyph list to be passed to <see cref="TextBuilder"/>.
+        /// </summary>
+        private readonly List<TextBuilderGlyph> charactersBacking = new List<TextBuilderGlyph>();
+
+        /// <summary>
+        /// The characters in local space.
+        /// </summary>
+        private List<TextBuilderGlyph> characters
+        {
+            get
+            {
+                computeCharacters();
+                return charactersBacking;
+            }
+        }
+
+        private bool isComputingCharacters;
+
+        /// <summary>
+        /// Compute character textures and positions.
+        /// </summary>
+        private void computeCharacters()
+        {
+            // Note : this feature can only use in osu-framework
+            // if (LoadState >= LoadState.Loaded)
+            //     ThreadSafety.EnsureUpdateThread();
+
+            if (store == null)
+                return;
+
+            if (charactersCache.IsValid)
+                return;
+
+            charactersBacking.Clear();
+
+            Debug.Assert(!isComputingCharacters, "Cyclic invocation of computeCharacters()!");
+            isComputingCharacters = true;
+
+            TextBuilder textBuilder = null;
+
+            try
+            {
+                if (string.IsNullOrEmpty(displayedText))
+                    return;
+
+                textBuilder = CreateExtraTextBuilder(store);
+                textBuilder.AddText(displayedText);
+            }
+            finally
+            {
+                if (requiresAutoSizedWidth)
+                    base.Width = (textBuilder?.Bounds.X ?? 0) + Padding.Right;
+                if (requiresAutoSizedHeight)
+                    base.Height = (textBuilder?.Bounds.Y ?? 0) + Padding.Bottom;
+
+                base.Width = Math.Min(base.Width, MaxWidth);
+
+                isComputingCharacters = false;
+                charactersCache.Validate();
+            }
+        }
+
+        private readonly LayoutValue parentScreenSpaceCache = new LayoutValue(Invalidation.DrawSize | Invalidation.Presence | Invalidation.DrawInfo, InvalidationSource.Parent);
+        private readonly LayoutValue localScreenSpaceCache = new LayoutValue(Invalidation.MiscGeometry, InvalidationSource.Self);
+
+        private readonly List<ScreenSpaceCharacterPart> screenSpaceCharactersBacking = new List<ScreenSpaceCharacterPart>();
+
+        /// <summary>
+        /// The characters in screen space. These are ready to be drawn.
+        /// </summary>
+        private List<ScreenSpaceCharacterPart> screenSpaceCharacters
+        {
+            get
+            {
+                computeScreenSpaceCharacters();
+                return screenSpaceCharactersBacking;
+            }
+        }
+
+        private void computeScreenSpaceCharacters()
+        {
+            if (!parentScreenSpaceCache.IsValid)
+            {
+                localScreenSpaceCache.Invalidate();
+                parentScreenSpaceCache.Validate();
+            }
+
+            if (localScreenSpaceCache.IsValid)
+                return;
+
+            screenSpaceCharactersBacking.Clear();
+
+            Vector2 inflationAmount = DrawInfo.MatrixInverse.ExtractScale().Xy;
+
+            foreach (var character in characters)
+            {
+                screenSpaceCharactersBacking.Add(new ScreenSpaceCharacterPart
+                {
+                    DrawQuad = ToScreenSpace(character.DrawRectangle.Inflate(inflationAmount)),
+                    InflationPercentage = Vector2.Divide(inflationAmount, character.DrawRectangle.Size),
+                    Texture = character.Texture
+                });
+            }
+
+            localScreenSpaceCache.Validate();
+        }
+
+        private readonly LayoutValue<Vector2> shadowOffsetCache = new LayoutValue<Vector2>(Invalidation.DrawInfo, InvalidationSource.Parent);
+
+        private Vector2 premultipliedShadowOffset =>
+            shadowOffsetCache.IsValid ? shadowOffsetCache.Value : shadowOffsetCache.Value = ToScreenSpace(shadowOffset * Font.Size) - ToScreenSpace(Vector2.Zero);
+
+        #endregion
+
+        #region Invalidation
+
+        private void invalidate(bool layout = false)
+        {
+            if (layout)
+                charactersCache.Invalidate();
+            parentScreenSpaceCache.Invalidate();
+            localScreenSpaceCache.Invalidate();
+
+            Invalidate(Invalidation.DrawNode);
+        }
+
+        #endregion
+
+        #region DrawNode
+
+        protected override DrawNode CreateDrawNode() => new LyricSpriteTextDrawNode(this);
+
+        #endregion
+
+        /// <summary>
+        /// The characters that should be excluded from fixed-width application. Defaults to (".", ",", ":", " ") if null.
+        /// </summary>
+        protected virtual char[] FixedWidthExcludeCharacters => null;
+
+        /// <summary>
+        /// The character to use to calculate the fixed width width. Defaults to 'm'.
+        /// </summary>
+        protected virtual char FixedWidthReferenceCharacter => 'm';
+
+        /// <summary>
+        /// The character to fallback to use if a character glyph lookup failed.
+        /// </summary>
+        protected virtual char FallbackCharacter => '?';
+
+        /// <summary>
+        /// Creates a <see cref="TextBuilder"/> to generate the character layout for this <see cref="LyricSpriteText"/>.
         /// </summary>
         /// <param name="store">The <see cref="ITexturedGlyphLookupStore"/> where characters should be retrieved from.</param>
         /// <returns>The <see cref="TextBuilder"/>.</returns>
-        protected override TextBuilder CreateTextBuilder(ITexturedGlyphLookupStore store)
+        protected virtual TextBuilder CreateTextBuilder(ITexturedGlyphLookupStore store)
         {
-            const int builder_max_width = int.MaxValue;
+            var excludeCharacters = FixedWidthExcludeCharacters ?? default_never_fixed_width_characters;
 
-            var excludeCharacters = FixedWidthExcludeCharacters;
+            float builderMaxWidth = requiresAutoSizedWidth
+                ? MaxWidth
+                : ApplyRelativeAxes(RelativeSizeAxes, new Vector2(Math.Min(MaxWidth, base.Width), base.Height), FillMode).X - Padding.Right;
 
-            // Calculate position
-            var rubyYPosition = Padding.Top;
-            var contentPosition = rubyYPosition + RubyFont.Size / 2 + RubyMargin;
-
-            // Print and save main texts
-            var charactersBacking = createMainTexts(Text, Font, contentPosition, Spacing);
-            Characters = charactersBacking.ToArray();
-
-            // Print ruby texts
-            createTexts(Rubies, RubyFont, rubyYPosition, RubySpacing);
-
-            // Calculate position and print romaji texts
-            var romajiYPosition = contentPosition + Characters.FirstOrDefault().Height + Characters.FirstOrDefault().YOffset + RomajiMargin;
-            createTexts(Romajies, RomajiFont, romajiYPosition, RomajiSpacing);
-
-            // Calculate position and return TextBuilder that do not renderer text anymore
-            var romajiTextSize = RomajiMargin + ((Romajies?.Any() ?? false) ? (charactersBacking.LastOrDefault().Height + charactersBacking.LastOrDefault().YOffset) : 0);
-            return new TextBuilder(store, Font, builder_max_width, UseFullGlyphHeight,
-                new Vector2(Padding.Left, contentPosition + romajiTextSize), Spacing, null,
-                excludeCharacters, FallbackCharacter);
-
-            // Create main text
-            List<TextBuilderGlyph> createMainTexts(string text, FontUsage font, float yPosition, Vector2 spacing)
+            if (AllowMultiline)
             {
-                var existCharacters = base.CreateTextBuilder(store).Characters;
-
-                var builder = new TextBuilder(store, Font, builder_max_width, UseFullGlyphHeight,
-                    new Vector2(Padding.Left, yPosition), spacing, existCharacters,
-                    excludeCharacters, FallbackCharacter);
-
-                builder.AddText(text);
-
-                return builder.Characters;
+                return new MultilineTextBuilder(store, Font, builderMaxWidth, UseFullGlyphHeight, new Vector2(Padding.Left, Padding.Top), Spacing, charactersBacking,
+                    excludeCharacters, FallbackCharacter, FixedWidthReferenceCharacter);
             }
 
-            // Create ruby and romaji texts
-            void createTexts(PositionText[] positionTexts, FontUsage font, float yPosition, Vector2 spacing)
+            if (Truncate)
             {
-                if (positionTexts != null)
-                {
-                    foreach (var positionText in positionTexts)
-                    {
-                        var text = positionText.Text;
-                        if (string.IsNullOrEmpty(text))
-                            continue;
+                return new TruncatingTextBuilder(store, Font, builderMaxWidth, ellipsisString, UseFullGlyphHeight, new Vector2(Padding.Left, Padding.Top), Spacing, charactersBacking,
+                    excludeCharacters, FallbackCharacter, FixedWidthReferenceCharacter);
+            }
 
-                        // Get text position
-                        var textPosition = new Vector2(getTextPosition(positionText, spacing.X), yPosition);
+            return new TextBuilder(store, Font, builderMaxWidth, UseFullGlyphHeight, new Vector2(Padding.Left, Padding.Top), Spacing, charactersBacking,
+                excludeCharacters, FallbackCharacter, FixedWidthReferenceCharacter);
+        }
 
-                        var builder = new TextBuilder(store, font, builder_max_width, UseFullGlyphHeight,
-                            textPosition, spacing, charactersBacking, excludeCharacters, FallbackCharacter);
+        public override string ToString() => $@"""{displayedText}"" " + base.ToString();
 
-                        builder.AddText(text);
-                    }
-                }
+        /// <summary>
+        /// Gets the base height of the font used by this text. If the font of this text is invalid, 0 is returned.
+        /// </summary>
+        public float LineBaseHeight
+        {
+            get
+            {
+                var baseHeight = store.GetBaseHeight(Font.FontName);
+                if (baseHeight.HasValue)
+                    return baseHeight.Value * Font.Size;
 
-                // Convert
-                float getTextPosition(PositionText text, float textSpacing)
-                {
-                    //It's magic number to let text in the center
-                    const float size_multiple = 1.4f;
-                    var centerPosition = (Characters[text.StartIndex].DrawRectangle.Left + Characters[text.EndIndex - 1].DrawRectangle.Right) / 2;
-                    var textWidth = text.Text.Sum(c => (store.Get(font.FontName, c)?.Width ?? 0) * font.Size * size_multiple) - (text.Text.Length) * textSpacing;
-                    return centerPosition - textWidth / 2;
-                }
+                if (string.IsNullOrEmpty(displayedText))
+                    return 0;
+
+                return store.GetBaseHeight(displayedText[0]).GetValueOrDefault() * Font.Size;
             }
         }
 
-        public float GetPercentageWidth(TextIndex startIndex, TextIndex endIndex, float percentage = 0)
-        {
-            if (Characters == null)
-                return 0;
-
-            var charLength = Characters.Length;
-            if (charLength == 0)
-                return 0;
-
-            startIndex = TextIndexUtils.Clamp(startIndex, 0, charLength - 1);
-            endIndex = TextIndexUtils.Clamp(endIndex, 0, charLength - 1);
-
-            var left = getWidth(startIndex);
-            var right = getWidth(endIndex);
-
-            var width = left * (1 - percentage) + right * percentage;
-            return width + Margin.Left;
-
-            float getWidth(TextIndex index)
-            {
-                switch (index.State)
-                {
-                    case TextIndex.IndexState.Start:
-                        return Characters[index.Index].DrawRectangle.Left;
-                    case TextIndex.IndexState.End:
-                        return Characters[index.Index].DrawRectangle.Right;
-                    default:
-                        throw new InvalidOperationException(nameof(index.State));
-                }
-            }
-        }
+        public IEnumerable<string> FilterTerms => displayedText.Yield();
     }
 }
